Extract custom user attributes from __init__
